LPS-77989 Move tests and resources
