+++ modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-context-helper-impl/src/test/java/com/liferay/portal/osgi/web/servlet/context/helper/internal/definition/MockFilter.java
@@ -12,20 +12,31 @@
  * details.
  */
 
-package com.liferay.portal.osgi.web.servlet.context.helper.definition;
+package com.liferay.portal.osgi.web.servlet.context.helper.internal.definition;
 
-import javax.servlet.GenericServlet;
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 
 /**
  * @author Miguel Pastor
  */
-public class MockServlet extends GenericServlet {
+public class MockFilter implements Filter {
 
 	@Override
-	public void service(
-		ServletRequest servletRequest, ServletResponse servletResponse) {
+	public void destroy() {
+	}
+
+	@Override
+	public void doFilter(
+		ServletRequest servletRequest, ServletResponse servletResponse,
+		FilterChain filterChain) {
+	}
+
+	@Override
+	public void init(FilterConfig filterConfig) {
 	}
 
 }
