+++ modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-context-helper-impl/src/test/resources/com/liferay/portal/osgi/web/servlet/context/helper/internal/definition/dependencies/custom-web-fragment-1.xml
@@ -12,11 +12,11 @@
 		<param-value>value1</param-value>
 	</context-param>
 	<listener>
-		<listener-class>com.liferay.portal.osgi.web.servlet.context.helper.definition.MockServletContextListener</listener-class>
+		<listener-class>com.liferay.portal.osgi.web.servlet.context.helper.internal.definition.MockServletContextListener</listener-class>
 	</listener>
 	<filter>
 		<filter-name>Mock Filter</filter-name>
-		<filter-class>com.liferay.portal.osgi.web.servlet.context.helper.definition.MockFilter</filter-class>
+		<filter-class>com.liferay.portal.osgi.web.servlet.context.helper.internal.definition.MockFilter</filter-class>
 	</filter>
 	<filter-mapping>
 		<filter-name>Mock Filter</filter-name>
