+++ modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-context-helper-impl/src/test/resources/com/liferay/portal/osgi/web/servlet/context/helper/internal/definition/dependencies/custom-web-fragment-circular-1.xml
