+++ modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-context-helper-impl/src/test/resources/com/liferay/portal/osgi/web/servlet/context/helper/internal/definition/dependencies/custom-web.xml
@@ -8,11 +8,11 @@
 	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_6.xsd"
 >
 	<listener>
-		<listener-class>com.liferay.portal.osgi.web.servlet.context.helper.definition.MockServletContextListener</listener-class>
+		<listener-class>com.liferay.portal.osgi.web.servlet.context.helper.internal.definition.MockServletContextListener</listener-class>
 	</listener>
 	<filter>
 		<filter-name>Mock Filter</filter-name>
-		<filter-class>com.liferay.portal.osgi.web.servlet.context.helper.definition.MockFilter</filter-class>
+		<filter-class>com.liferay.portal.osgi.web.servlet.context.helper.internal.definition.MockFilter</filter-class>
 	</filter>
 	<filter-mapping>
 		<filter-name>Mock Filter</filter-name>
@@ -20,7 +20,7 @@
 	</filter-mapping>
 	<servlet>
 		<servlet-name>Mock Servlet</servlet-name>
-		<servlet-class>com.liferay.portal.osgi.web.servlet.context.helper.definition.MockServlet</servlet-class>
+		<servlet-class>com.liferay.portal.osgi.web.servlet.context.helper.internal.definition.MockServlet</servlet-class>
 		<load-on-startup>0</load-on-startup>
 	</servlet>
 	<servlet-mapping>
